+++ firmware/font/hextokff.py
@@ -28,7 +28,6 @@ import math
 import os
 import json
 
-
 BYTE_LEN = 2
 CHAR_LIST_EXCEPT_ASIAN = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789 !#$%&'()*+,-./:;<=>?@`[\\]^_\"{|}~█₿ ⊚↳«»…"
 DEFAULT_CODEPOINTS = [ord(char) for char in CHAR_LIST_EXCEPT_ASIAN]